feat: divide files in processes

+++ apps/server/src/services/processService.ts
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 
-import { unlinkSync, renameSync } from 'fs';
+import { unlinkSync, renameSync, existsSync } from 'fs';
 
 import { IAddProcessResponse, IStatusResponse, AddProcess, IProcessedFiles, NewMessage } from '@lib/types';
 
@@ -248,11 +248,13 @@ export const completeById = (processId: string): IStatusResponse => {
 
       if (toPath && requestFN) {
         try {
-          renameSync(getPathMessages(process, requestFN), toPath);
+          if (existsSync(getPathMessages(process, requestFN))) {
+            renameSync(getPathMessages(process, requestFN), toPath);
 
-          const i = process.files.indexOf(requestFN);
-          process.files.splice(i, 1);
-          saveProcessList();
+            const i = process.files.indexOf(requestFN);
+            process.files.splice(i, 1);
+            saveProcessList();
+          }
         } catch (err) {
           error = `Robust-protocol.completeProcess: не удалось перенести файл ${requestFN} из MESSAGES,
             ${err instanceof Error ? err.message : 'ошибка'}`;
